index.css -> globals.css 로 내용 다 옮겼습니다. index.css 파일 삭제 이미지 호스팅(img 불러오는 방법 바꿈)
