--- src/app/index.css
@@ -1,73 +1,0 @@
-/* 내비게이션바-로그아웃 버튼 배경 효과 */
-.hover-bg-gradient:hover {
-    background-image: linear-gradient(to right, black, rgb(71, 9, 71));
-}
-
-.hover-bg-gradient-LoginBtn:hover {
-    background-color: rgb(112, 65, 156);
-}
-
-/* 로그인페이지- 로그인 버튼들 효과 */
-.fade-in-box {
-    animation: fadein 3s;
-  }
-  @keyframes fadein {
-      from {
-          opacity: 0;
-      }
-      to {
-          opacity: 1;
-      }
-  }
-
-  .LogInBtns {
-    background: transparent;
-    color: rgba(207, 124, 255, 0.7);
-    transition: all 0.7s ease-out;
-    background: linear-gradient(270deg, rgba(167, 106, 223, 0.8), rgba(93, 52, 146, 0.8), rgba(34, 34, 34, 0), rgba(34, 34, 34, 0));
-    background-position: 1% 50%;
-    background-size: 300% 300%;
-    text-decoration: none;
-    border: 3px solid rgba(185, 101, 235, 0.3);
-  }
-  
-  .LogInBtns:hover {
-    color: #fff;
-    border: 1px solid rgba(223,190,106,0);
-    color: white;
-    background-position: 99% 50%;
-    cursor: pointer;
-  }
-
-  /* 로그인 페이지-배경 효과 */
-
-  .wave {
-    opacity: .4;
-    position: fixed;
-    top: 0;
-    left: 0;
-    background: rgba(111, 0, 255, 0.164);
-    width: 100%;
-    height: 100%;
-    margin-left: -20%;
-    margin-top: -20%;
-    transform-origin: 50% 48%;
-    border-radius: 100%;
-    animation: drift 3000ms infinite linear;
-    z-index:1;
-  }
-  
-  .wave.-three {
-    animation: drift 5000ms infinite linear;
-  }
-  
-  .wave.-two {
-    animation: drift 7000ms infinite linear;
-    opacity: .1;
-    background: rgba(187, 255, 0, 0.603);
-  }
-  
-  @keyframes drift {
-    from { transform: rotate(0deg); }
-    from { transform: rotate(360deg); }
-  }
