+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable import/extensions */
 // import MusicBar from './components/Musicbar/Musicbar';
 // import NavBar from './components/NavBar/NavigationBar';
-import LogIn from './LoginPage';
+import LogIn from './LogInPage/page';
 // App 컴포넌트
 export default function Home() {
   return (
